refactor: move around canvas related files
