+++ src/events/index.js
@@ -1,4 +1,3 @@
-export * from "./CanvasEvents.js"
 export * from "./EventEmitter.js"
 
 // MARK: - keyboard
