+++ src/graphics/canvas/CanvasEvents.js
